refactor(partners): migrate partner declarations + directory into modules/
