+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/directory/navigation-menu-items/partner-content.navigation-menu-item.ts
