+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/directory/navigation-menu-items/partner-workspace-folder.navigation-menu-item.ts
