+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/directory/navigation-menu-items/partners-validated.navigation-menu-item.ts
@@ -1,6 +1,6 @@
 import { NavigationMenuItemType, defineNavigationMenuItem } from 'twenty-sdk/define';
 
-import { PARTNERS_VALIDATED_VIEW_UNIVERSAL_IDENTIFIER } from 'src/views/partners-validated.view';
+import { PARTNERS_VALIDATED_VIEW_UNIVERSAL_IDENTIFIER } from 'src/modules/partner/directory/views/partners-validated.view';
 
 export default defineNavigationMenuItem({
   universalIdentifier: 'c9cbe1b6-7c07-4327-bd7a-018d0ed0db33',
