+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/directory/page-layouts/partner-link.page-layout.ts
@@ -1,7 +1,7 @@
 import { PageLayoutTabLayoutMode, definePageLayout } from 'twenty-sdk/define';
 
 import { PARTNER_LINK_OBJECT_UNIVERSAL_IDENTIFIER } from 'src/constants/universal-identifiers';
-import { PARTNER_LINK_RECORD_PAGE_FIELDS_VIEW_ID } from 'src/views/partner-link-record-page-fields.view';
+import { PARTNER_LINK_RECORD_PAGE_FIELDS_VIEW_ID } from 'src/modules/partner/directory/views/partner-link-record-page-fields.view';
 
 export default definePageLayout({
   universalIdentifier: 'e2b44276-babe-4f29-a94d-8fa220fe6483',
