+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/companies-as-partner-user-on-workspace-member.field.ts
