+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-company.field.ts
