+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-content-customer-company.field.ts
