+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-content-customer-person.field.ts
