+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-content-partner.field.ts
