+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-contents-as-partner-user-on-workspace-member.field.ts
