+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-contents-on-company.field.ts
