+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-profiles-as-user-on-workspace-member.field.ts
