+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-services-on-partner.field.ts
