+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/fields/partner-user-on-person.field.ts
