+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/objects/partner-content.object.ts
