+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/objects/partner-link.object.ts
