+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/objects/partner-service.object.ts
