+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/objects/partner.object.ts
