+++ packages/twenty-apps/internal/twenty-partners/src/navigation-menu-items/my-case-studies.navigation-menu-item.ts
@@ -3,7 +3,7 @@ import { NavigationMenuItemType, defineNavigationMenuItem } from 'twenty-sdk/def
 import { MY_CASE_STUDIES_NAV_UNIVERSAL_IDENTIFIER } from 'src/constants/universal-identifiers';
 
 import { MY_CASE_STUDIES_PAGE_LAYOUT_ID } from 'src/constants/my-case-studies.constants';
-import { PARTNER_WORKSPACE_FOLDER_UNIVERSAL_IDENTIFIER } from './partner-workspace-folder.navigation-menu-item';
+import { PARTNER_WORKSPACE_FOLDER_UNIVERSAL_IDENTIFIER } from 'src/modules/partner/directory/navigation-menu-items/partner-workspace-folder.navigation-menu-item';
 
 export default defineNavigationMenuItem({
   universalIdentifier: MY_CASE_STUDIES_NAV_UNIVERSAL_IDENTIFIER,
